requirements.txt
fastapi==0.112.0
pydantic==2.8.2
uvicorn==0.30.6
pytest==8.3.2
openai==1.40.6
sqlalchemy==2.0.32
python-multipart==0.0.9
websocket-client==1.8.0
